Múa thêm ở FE

+++ backend/src/main/java/com/travel_agent/controllers/UserController.java
@@ -11,6 +11,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.security.access.prepost.PreAuthorize;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -22,6 +23,7 @@ public class UserController {
     private final UserService userService;
 
     @PostMapping
+    @PreAuthorize("hasRole('GUEST')")
     public ResponseEntity<ResponseObject> createUser(@RequestBody UserDTO userDTO) {
         String username = userDTO.getUsername();
         String password = userDTO.getPassword();
@@ -46,7 +48,15 @@ public class UserController {
     }
 
     @GetMapping("/")
+    @PreAuthorize("hasRole('USER')")
     public ResponseEntity<ResponseObject> getUserById(@CurrentUserId Integer userId) {
+        if (userId == null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseObject.builder()
+                    .message("User has not logged in!")
+                    .responseCode(HttpStatus.BAD_REQUEST.value())
+                    .build());
+        }
+
         UserDTO user = userService.getUserById(userId);
         return ResponseEntity.ok(ResponseObject.builder()
                 .message("User retrieved successfully")
@@ -56,7 +66,15 @@ public class UserController {
     }
 
     @PutMapping("/")
+    @PreAuthorize("hasRole('USER')")
     public ResponseEntity<ResponseObject> updateUser(@CurrentUserId Integer userId, @RequestBody UserDTO userDTO) throws ReflectionException{
+        if (userId == null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseObject.builder()
+                    .message("User has not logged in!")
+                    .responseCode(HttpStatus.BAD_REQUEST.value())
+                    .build());
+        }
+        
         UserDTO updatedUser = userService.updateUser(userId, userDTO);
         return ResponseEntity.ok(ResponseObject.builder()
                 .message("User updated successfully")
@@ -66,7 +84,14 @@ public class UserController {
     }
 
     @DeleteMapping("/")
+    @PreAuthorize("hasRole('USER')")
     public ResponseEntity<ResponseObject> deleteUser(@CurrentUserId Integer userId) {
+        if (userId == null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ResponseObject.builder()
+                    .message("User has not logged in!")
+                    .responseCode(HttpStatus.BAD_REQUEST.value())
+                    .build());
+        }
         userService.deleteUser(userId);
         return ResponseEntity.ok(ResponseObject.builder()
                 .message("User deleted successfully")
@@ -74,13 +99,14 @@ public class UserController {
                 .build());
     }
 
-    @GetMapping
-    public ResponseEntity<ResponseObject> getAllUsers() {
-        List<UserDTO> users = userService.getAllUsers();
-        return ResponseEntity.ok(ResponseObject.builder()
-                .message("List of users retrieved successfully")
-                .data(users)
-                .responseCode(HttpStatus.OK.value())
-                .build());
-    }
+    // @GetMapping
+    // @PreAuthorize("hasRole('ADMIN')")
+    // public ResponseEntity<ResponseObject> getAllUsers() {
+    //     List<UserDTO> users = userService.getAllUsers();
+    //     return ResponseEntity.ok(ResponseObject.builder()
+    //             .message("List of users retrieved successfully")
+    //             .data(users)
+    //             .responseCode(HttpStatus.OK.value())
+    //             .build());
+    // }
 }
